moved Clients from model/utils -> model/
