+++ src/main/kotlin/com/redhat/devtools/intellij/kubernetes/model/resource/openshift/BuildsOperator.kt
@@ -13,7 +13,7 @@ package com.redhat.devtools.intellij.kubernetes.model.resource.openshift
 import com.redhat.devtools.intellij.kubernetes.model.resource.NamespacedOperation
 import com.redhat.devtools.intellij.kubernetes.model.resource.NamespacedResourceOperator
 import com.redhat.devtools.intellij.kubernetes.model.resource.ResourceKind
-import com.redhat.devtools.intellij.kubernetes.model.util.Clients
+import com.redhat.devtools.intellij.kubernetes.model.Clients
 import io.fabric8.openshift.api.model.Build
 import io.fabric8.openshift.client.OpenShiftClient
